some more work.
The enum method for getting the new review items should work without problems (besides timing issues).

However, the code to parse a review item needs to be fixed. I can copy most of the code in the ReviewItems class in SOCRV.Net and modify it a bit to fix my needs.

If I really want to go overboard, I would make this whole process an external program so that multiple requests can't be going at once (and increase the chances that SO blocks my IP).